feat(admin): reorganize overview activity charts
