+++ admin/src/app/(admin)/earn/rebalance/rebalance-monitor-client.tsx
@@ -41,11 +41,6 @@ import type {
 } from "@/lib/kamino/timescale-reserve-monitor.shared";
 
 import { SafeReserveApyChart } from "../safe-reserve-apy-chart";
-import {
-  AutodepositActivityChart,
-  type SerializedAutodepositTimeSeriesRange,
-} from "./autodeposit-activity-chart";
-import { OptimizationVolumeChart } from "./optimization-volume-chart";
 import { PreviousMonthRebalanceChart } from "./previous-month-rebalance-chart";
 import { RebalanceActivityChart } from "./rebalance-activity-chart";
 import type {
@@ -80,13 +75,6 @@ type SerializedRebalanceDecisionRow = Omit<
   confirmedSlot: string | null;
 };
 
-type SerializedOptimizationVolumePoint = {
-  confirmedCount: number;
-  cumulativeAmountRaw: string;
-  dailyAmountRaw: string;
-  date: string;
-};
-
 type SerializedRebalanceAuditRow = Omit<
   RebalanceAuditRow,
   "amountRaw" | "confirmedSlot" | "submittedSlot"
@@ -99,9 +87,7 @@ type SerializedRebalanceAuditRow = Omit<
 type RebalanceApiData = {
   activity: RebalanceActivityPoint[];
   apyData: SafeReserveApyMonitorData;
-  autodeposit: SerializedAutodepositTimeSeriesRange[];
   decisions: SerializedRebalanceDecisionRow[];
-  optimizationVolume: SerializedOptimizationVolumePoint[];
   previousMonthRebalances: PreviousMonthRebalancePoint[];
   routes: SerializedActiveReserveRouteRow[];
 };
@@ -1286,8 +1272,6 @@ function RebalanceMonitorFallback() {
       <CurrentReserveApyFallback />
       <ApyChartFallback />
       <ApyChartFallback />
-      <ApyChartFallback />
-      <ApyChartFallback />
       <RebalanceDecisionAuditFallback />
     </div>
   );
@@ -1365,8 +1349,6 @@ export function RebalanceMonitorClient() {
       />
       <RebalanceActivityChart data={state.data.activity} />
       <PreviousMonthRebalanceChart data={state.data.previousMonthRebalances} />
-      <AutodepositActivityChart data={state.data.autodeposit} />
-      <OptimizationVolumeChart data={state.data.optimizationVolume} />
       <RebalanceAuditCard reserveLabels={reserveLabels} />
     </div>
   );
